[TS-418] Refactor: social login change to custom hook
